keeping update for better welcomeview page and consistent formatting
