+++ BravoBall.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -6,8 +6,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/rive-app/rive-ios",
       "state" : {
-        "revision" : "4a1795829a2291a722fdfb71f05424146413d759",
-        "version" : "6.1.3"
+        "revision" : "51f1bcdfce0d956cab2994c64e29b5fd3e7c171e",
+        "version" : "6.3.1"
       }
     }
   ],
