+++ Tekk/Features/Onboarding/Views/WelcomeQuestions.swift
@@ -30,8 +30,8 @@ struct WelcomeQuestions: View {
     @State private var isEditingLastName = false
     
     var body: some View {
-        VStack(spacing: 25) {
-            // first name, Zstack so placeholder is on top of input text
+        VStack(spacing: 15) {
+            // first name
             ZStack(alignment: .leading) {
                 if firstName.isEmpty {
                     Text("First Name")
@@ -40,18 +40,19 @@ struct WelcomeQuestions: View {
                         .font(.custom("Poppins-Bold", size: 16))
                 }
                 TextField("", text: $firstName)
-                .padding()
-                .foregroundColor(globalSettings.primaryDarkColor)
-                .background(Color.clear)
-                .overlay(
-                    RoundedRectangle(cornerRadius: 20)
-                        .stroke(Color.gray, lineWidth: 1)
-                )
-                .font(.custom("Poppins-Bold", size: 16))
+                    .padding()
+                    .foregroundColor(globalSettings.primaryDarkColor)
+                    .background(Color.clear)
+                    .overlay(
+                        RoundedRectangle(cornerRadius: 20)
+                            .stroke(Color.gray, lineWidth: 1)
+                    )
+                    .font(.custom("Poppins-Bold", size: 16))
             }
             .frame(height: 60)
+            .padding(.horizontal, 20)
             
-            // last name, Zstack so placeholder is on top of input text
+            // last name (same modifications as first name)
             ZStack(alignment: .leading) {
                 if lastName.isEmpty {
                     Text("Last Name")
@@ -60,27 +61,24 @@ struct WelcomeQuestions: View {
                         .font(.custom("Poppins-Bold", size: 16))
                 }
                 TextField("", text: $lastName)
-                .padding()
-                .foregroundColor(globalSettings.primaryDarkColor)
-                .background(Color.clear)
-                .overlay(
-                    RoundedRectangle(cornerRadius: 20)
-                        .stroke(Color.gray, lineWidth: 1)
-                )
-                .font(.custom("Poppins-Bold", size: 16))
+                    .padding()
+                    .foregroundColor(globalSettings.primaryDarkColor)
+                    .background(Color.clear)
+                    .overlay(
+                        RoundedRectangle(cornerRadius: 20)
+                            .stroke(Color.gray, lineWidth: 1)
+                    )
+                    .font(.custom("Poppins-Bold", size: 16))
             }
             .frame(height: 60)
+            .padding(.horizontal, 20)
             
-            // confined version of structure drop down menus
-            dropDownMenu(title: $selectedAge, options: ageOptions, placeholder: "Select your Age")
-            dropDownMenu(title: $selectedLevel, options: levelOptions, placeholder: "Select your Level")
-            dropDownMenu(title: $selectedPosition, options: positionOptions, placeholder: "Select your Position")
-                .frame(height: 60)
-                .padding(.bottom, 325)
-                .transition(.move(edge: .bottom))
+            // dropdown menus
+            DropdownMenu(title: $selectedAge, options: ageOptions, placeholder: "Select your Age")
+            DropdownMenu(title: $selectedLevel, options: levelOptions, placeholder: "Select your Level")
+            DropdownMenu(title: $selectedPosition, options: positionOptions, placeholder: "Select your Position")
         }
-        .padding(.horizontal)
-        .padding(.top, 500)
+        .padding(.top, 20)
     }
 }
